refactor-3 Pass 3C-2: extract SpawnControlUI from ParsekUI
Move spawn control window into dedicated UI/SpawnControlUI.cs class.
14 fields, SpawnSortColumn enum, and 7 methods moved. ParsekUI delegates
via spawnControlUI.DrawIfOpen() and toggle via IsOpen property.
